# fs16-backend/swagger-definition.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        image:
          type: string
      required:
        - _id
        - name
        - image
      example:
        _id: "1"
        name: "Category Name"
        image: "https://example.com/category_image.jpg"
...
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: string
        image:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        stock:
          type: number
        size:
          $ref: "#/components/schemas/Size"
      required:
        - _id
        - name
        - price
        - description
        - image
        - category
        - stock
      example:
        _id: "1"
        name: "Product Name"
        price: 29.99
        description: "Product Description"
        image: "https://example.com/product_image.jpg"
        category:
          _id: "category_id"
          name: "Category Name"
          # ... other properties of the Category schema
        stock: 100
        size:
          _id: "size_id"
          name: "Size Name"
          # ... other properties of the Size schema
    Size:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
      required:
        - _id
        - name
      example:
        _id: "1"
        name: "Size Name"
...
paths:
...
  /products/{productId}:
    get:
      summary: Get details of a product
      tags:
        - Product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
          example: "productId"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not Found
          content:
            application/json:
              example:
                message: "Not Found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                message: "Internal server error"
    put:
      summary: Update a product
      tags:
        - Product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
          example: "productId"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                  minimum: 0
                description:
                  type: string
                category:
                  type: string
                stock:
                  type: number
                size:
                  type: string
                image:
                  type: string
              required:
                - name
                - price
                - description
                - category
                - stock
                - size
                - image
      responses:
        "200":
          description: Successful update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                message: "Unauthorized"
        "404":
          description: Not Found
          content:
            application/json:
              example:
                message: "Not Found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                message: "Internal server error"
    delete:
      summary: Delete a product
      tags:
        - Product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
          example: "productId"
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Successful deletion
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                message: "Unauthorized"
        "404":
          description: Not Found
          content:
            application/json:
              example:
                message: "Not Found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                message: "Internal server error"
...